renamed resources to modules
