+++ packages/ssx-sdk/tests/modules/UserAuthorization.test.ts
@@ -3,7 +3,7 @@ import { TextEncoder as TE, TextDecoder as TD } from 'util';
 global.TextEncoder = TE;
 global.TextDecoder = TD;
 
-import { UserAuthorization } from '../../resources';
+import { UserAuthorization } from '../../src/modules';
 
 const testingUtils = generateTestingUtils({ providerType: 'MetaMask' });
